+++ test/contrib/test_pyopenssl.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 import unittest
 
-from nose.plugins.skip import SkipTest
 import pytest
 
 try:
@@ -9,7 +8,7 @@ try:
                                            extract_from_urllib3,
                                            _dnsname_to_stdlib)
 except ImportError as e:
-    raise SkipTest('Could not import PyOpenSSL: %r' % e)
+    pytestmark = pytest.mark.skip('Could not import PyOpenSSL: %r' % e)
 
 
 from ..with_dummyserver.test_https import TestHTTPS, TestHTTPS_TLSv1  # noqa: F401