+++ test/contrib/test_securetransport.py
@@ -1,13 +1,13 @@
 # -*- coding: utf-8 -*-
-from nose.plugins.skip import SkipTest
+import pytest
 
 try:
     from urllib3.contrib.securetransport import (inject_into_urllib3,
                                                  extract_from_urllib3)
 except ImportError as e:
-    raise SkipTest('Could not import SecureTransport: %r' % e)
+    pytestmark = pytest.mark.skip('Could not import SecureTransport: %r' % e)
 
-raise SkipTest('SecureTransport currently not supported in v2!')
+pytestmark = pytest.mark.skip('SecureTransport currently not supported on v2!')
 
 from ..with_dummyserver.test_https import TestHTTPS, TestHTTPS_TLSv1  # noqa: F401
 from ..with_dummyserver.test_socketlevel import (  # noqa: F401