+++ test/test_connection.py
@@ -1,31 +1,26 @@
 import datetime
 import mock
-import sys
+
+import pytest
 
 from urllib3._sync.connection import RECENT_DATE
 from urllib3.util.ssl_ import CertificateError, match_hostname
 
-import pytest
-
-if sys.version_info >= (2, 7):
-    import unittest
-else:
-    import unittest2 as unittest
-
-
-class TestConnection(unittest.TestCase):
+class TestConnection(object):
     """
     Tests in this suite should not make any network requests or connections.
     """
     def test_match_hostname_no_cert(self):
         cert = None
         asserted_hostname = 'foo'
-        self.assertRaises(ValueError, match_hostname, cert, asserted_hostname)
+        with pytest.raises(ValueError):
+            match_hostname(cert, asserted_hostname)
 
     def test_match_hostname_empty_cert(self):
         cert = {}
         asserted_hostname = 'foo'
-        self.assertRaises(ValueError, match_hostname, cert, asserted_hostname)
+        with pytest.raises(ValueError):
+            match_hostname(cert, asserted_hostname)
 
     def test_match_hostname_match(self):
         cert = {'subjectAltName': [('DNS', 'foo')]}
@@ -39,13 +34,13 @@ class TestConnection(unittest.TestCase):
             with mock.patch('urllib3.util.ssl_.log.error') as mock_log:
                 match_hostname(cert, asserted_hostname)
         except CertificateError as e:
-            self.assertEqual(str(e), "hostname 'bar' doesn't match 'foo'")
+            assert str(e) == "hostname 'bar' doesn't match 'foo'"
             mock_log.assert_called_once_with(
                 'Certificate did not match expected hostname: %s. '
                 'Certificate: %s',
                 'bar', {'subjectAltName': [('DNS', 'foo')]}
             )
-            self.assertEqual(e._peer_cert, cert)
+            assert e._peer_cert == cert
 
     @pytest.mark.xfail
     def test_recent_date(self):
@@ -54,8 +49,4 @@ class TestConnection(unittest.TestCase):
         # When this test fails update urllib3.connection.RECENT_DATE
         # according to the rules defined in that file.
         two_years = datetime.timedelta(days=365 * 2)
-        self.assertGreater(RECENT_DATE, (datetime.datetime.today() - two_years).date())
-
-
-if __name__ == '__main__':
-    unittest.main()
+        assert RECENT_DATE > (datetime.datetime.today() - two_years).date()