+++ test/with_dummyserver/test_connectionpool.py
@@ -180,12 +180,14 @@ class TestConnectionPoolTimeouts(SocketDummyServerTestCase):
         pool = HTTPConnectionPool(host, port, timeout=timeout)
         self.addCleanup(pool.close)
         conn = pool._get_conn()
+        self.addCleanup(conn.close)
         self.assertRaises(ConnectTimeoutError, pool._make_request, conn, 'GET', '/')
 
         timeout = Timeout(connect=3, read=5, total=SHORT_TIMEOUT)
         pool = HTTPConnectionPool(host, port, timeout=timeout)
         self.addCleanup(pool.close)
         conn = pool._get_conn()
+        self.addCleanup(conn.close)
         self.assertRaises(ConnectTimeoutError, pool._make_request, conn, 'GET', '/')
 
     def test_total_timeout(self):
@@ -307,7 +309,9 @@ class TestConnectionPool(HTTPDummyServerTestCase):
         # This test needs to be here in order to be run. socket.create_connection actually tries
         # to connect to the host provided so we need a dummyserver to be running.
         pool = HTTPConnectionPool(self.host, self.port)
+        self.addCleanup(pool.close)
         conn = pool._get_conn()
+        self.addCleanup(conn.close)
         pool._make_request(conn, 'GET', '/')
         tcp_nodelay_setting = conn._sock.getsockopt(socket.IPPROTO_TCP, socket.TCP_NODELAY)
         self.assertTrue(tcp_nodelay_setting)
@@ -343,12 +347,15 @@ class TestConnectionPool(HTTPDummyServerTestCase):
         # This test needs to be here in order to be run. socket.create_connection actually tries
         # to connect to the host provided so we need a dummyserver to be running.
         pool = HTTPConnectionPool(self.host, self.port)
+        self.addCleanup(pool.close)
         # Get the HTTPConnection instance
         conn = pool._new_conn()
+        self.addCleanup(conn.close)
         # Update the default socket options
         conn.default_socket_options += [(socket.SOL_SOCKET, socket.SO_KEEPALIVE, 1)]
         conn.connect()
         s = conn._sock
+        self.addCleanup(s.close)
         nagle_disabled = s.getsockopt(socket.IPPROTO_TCP, socket.TCP_NODELAY) > 0
         using_keepalive = s.getsockopt(socket.SOL_SOCKET, socket.SO_KEEPALIVE) > 0
         self.assertTrue(nagle_disabled)
@@ -367,17 +374,28 @@ class TestConnectionPool(HTTPDummyServerTestCase):
     def test_timeout_success(self):
         timeout = Timeout(connect=3, read=5, total=None)
         pool = HTTPConnectionPool(self.host, self.port, timeout=timeout)
+        self.addCleanup(pool.close)
         pool.request('GET', '/')
         # This should not raise a "Timeout already started" error
         pool.request('GET', '/')
 
         pool = HTTPConnectionPool(self.host, self.port, timeout=timeout)
+        self.addCleanup(pool.close)
         # This should also not raise a "Timeout already started" error
         pool.request('GET', '/')
 
         timeout = Timeout(total=None)
         pool = HTTPConnectionPool(self.host, self.port, timeout=timeout)
+        self.addCleanup(pool.close)
         pool.request('GET', '/')
+
+    def test_redirect(self):
+        r = self.pool.request('GET', '/redirect', fields={'target': '/'}, redirect=False)
+        self.assertEqual(r.status, 303)
+
+        r = self.pool.request('GET', '/redirect', fields={'target': '/'})
+        self.assertEqual(r.status, 200)
+        self.assertEqual(r.data, b'Dummy server!')
 
     def test_bad_connect(self):
         pool = HTTPConnectionPool('badhost.invalid', self.port)
@@ -389,6 +407,7 @@ class TestConnectionPool(HTTPDummyServerTestCase):
 
     def test_keepalive(self):
         pool = HTTPConnectionPool(self.host, self.port, block=True, maxsize=1)
+        self.addCleanup(pool.close)
 
         r = pool.request('GET', '/keepalive?close=0')
         r = pool.request('GET', '/keepalive?close=0')
@@ -400,6 +419,7 @@ class TestConnectionPool(HTTPDummyServerTestCase):
     def test_keepalive_close(self):
         pool = HTTPConnectionPool(self.host, self.port,
                                   block=True, maxsize=1, timeout=2)
+        self.addCleanup(pool.close)
 
         r = pool.request('GET', '/keepalive?close=1', retries=0,
                          headers={
@@ -499,6 +519,7 @@ class TestConnectionPool(HTTPDummyServerTestCase):
 
     def test_connection_count(self):
         pool = HTTPConnectionPool(self.host, self.port, maxsize=1)
+        self.addCleanup(pool.close)
 
         pool.request('GET', '/')
         pool.request('GET', '/')
@@ -509,6 +530,7 @@ class TestConnectionPool(HTTPDummyServerTestCase):
 
     def test_connection_count_bigpool(self):
         http_pool = HTTPConnectionPool(self.host, self.port, maxsize=16)
+        self.addCleanup(http_pool.close)
 
         http_pool.request('GET', '/')
         http_pool.request('GET', '/')
@@ -519,6 +541,7 @@ class TestConnectionPool(HTTPDummyServerTestCase):
 
     def test_partial_response(self):
         pool = HTTPConnectionPool(self.host, self.port, maxsize=1)
+        self.addCleanup(pool.close)
 
         req_data = {'lol': 'cat'}
         resp_data = urlencode(req_data).encode('utf-8')
@@ -580,6 +603,7 @@ class TestConnectionPool(HTTPDummyServerTestCase):
 
         # Check default state
         pool = HTTPConnectionPool(self.host, self.port, maxsize=MAXSIZE)
+        self.addCleanup(pool.close)
         self.assertEqual(pool.num_connections, 0)
         self.assertEqual(pool.pool.qsize(), MAXSIZE)
 
@@ -625,6 +649,7 @@ class TestConnectionPool(HTTPDummyServerTestCase):
                 continue
             pool = HTTPConnectionPool(self.host, self.port,
                                       source_address=addr, retries=False)
+            self.addCleanup(pool.close)
             r = pool.request('GET', '/source_address')
             self.assertEqual(r.data, b(addr[0]))
 
@@ -667,6 +692,7 @@ class TestConnectionPool(HTTPDummyServerTestCase):
 
     def test_mixed_case_hostname(self):
         pool = HTTPConnectionPool("LoCaLhOsT", self.port)
+        self.addCleanup(pool.close)
         response = pool.request('GET', "http://LoCaLhOsT:%d/" % self.port)
         self.assertEqual(response.status, 200)